created core grav module; revised index.js so that helper modules and components are exposed; revised folder tree + fixed breaking changes in grav.test example
