+++ examples/grav.test.js
@@ -1,7 +1,7 @@
-const Grav = require('../index');
 const creds = require('../creds');
-const AckParser = require('../core/parsers/ack.parser');
-const ParseContext = require('../core/parsers/_parse.context');
+const { 
+  Grav, AckParser, ParseContext 
+} = require('../index');
 
 const ackParser = new AckParser();
 const context = new ParseContext(ackParser);
